17-pallet-vision-drone-construction/requirements.txt
streamlit>=1.30
pandas>=2.1
pillow>=10.0
openai>=1.40
python-dotenv>=1.0
